voice-core/pyproject.toml
[tool.poetry]
name = "voice-core"
version = "0.1.0"
description = "Voice Core (Layer 1) - Immutable voice AI foundation using Pipecat"
authors = ["SpotFunnel AI"]
readme = "README.md"
requires-python = ">=3.11"

[tool.poetry.dependencies]
python = "^3.11"
pipecat-ai = {extras = ["daily", "deepgram", "openai", "elevenlabs"], version = "^0.0.46"}
python-dotenv = "^1.0.1"
aiohttp = "^3.9.1"
pytz = "^2024.1"
grpcio = "^1.62.0"
grpcio-tools = "^1.62.0"
protobuf = "^4.25.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.4"
pytest-mock = "^3.12.0"
black = "^24.0.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
